+++ src/api/client.ts
@@ -11,7 +11,6 @@ export interface ApiClient {
   uploadObject: (object: Record<string, unknown>) => Promise<string>;
   startBulkUpload: (urls: string[]) => Promise<UrlToJobId>;
   checkBulkUploadJob: (urls: string[]) => Promise<UploadJobStatus>;
-  generateDefaultMetadata: (name: string) => Promise<string>;
 }
 
 export const createClient = (apiUri: string): ApiClient => {
@@ -40,10 +39,6 @@ export const createClient = (apiUri: string): ApiClient => {
       logger.debug(`Check bulk upload status for ${jobIds.length} jobs`);
       const status = actions.checkBulkUploadJob(apiUri, jobIds);
       return status;
-    },
-    generateDefaultMetadata: (name: string) => {
-      logger.debug(`Generating metadata for domain ${name}`);
-      return actions.generateDefaultMetadata(apiUri, name);
     },
   };
